requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-functions==1.21.3
beautifulsoup4==4.12.3
certifi==2024.8.30
distro==1.9.0
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.1
lxml==5.3.0
numpy==2.1.3
openai==1.54.4
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pymssql==2.3.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
